SELC-2244 - refactor code (#555)
